+++ src/metatrain/pet/model.py
@@ -5,21 +5,21 @@ from typing import Any, Dict, List, Literal, Optional
 import metatensor.torch
 import torch
 from metatensor.torch import Labels, TensorBlock, TensorMap
-from metatensor.torch.atomistic import (
-    MetatensorAtomisticModel,
+from metatomic.torch import (
+    AtomisticModel,
     ModelCapabilities,
     ModelMetadata,
     ModelOutput,
     NeighborListOptions,
     System,
 )
-from torch import nn
 
+from metatrain.utils.abc import ModelInterface
 from metatrain.utils.additive import ZBL, CompositionModel
 from metatrain.utils.data import DatasetInfo, TargetInfo
 from metatrain.utils.dtype import dtype_to_str
 from metatrain.utils.long_range import DummyLongRangeFeaturizer, LongRangeFeaturizer
-from metatrain.utils.metadata import append_metadata_references
+from metatrain.utils.metadata import merge_metadata
 from metatrain.utils.scaler import Scaler
 from metatrain.utils.sum_over_atoms import sum_over_atoms
 
@@ -29,9 +29,9 @@ from .modules.transformer import CartesianTransformer
 from .modules.utilities import cutoff_func
 
 
-class PET(torch.nn.Module):
+class PET(ModelInterface):
     """
-    Native metatrain implementation of the PET architecture.
+    Metatrain-native implementation of the PET architecture.
 
     Originally proposed in work (https://arxiv.org/abs/2305.19302v3),
     and published in the `pet` package (https://github.com/spozdn/pet).
@@ -678,6 +678,7 @@ class PET(torch.nn.Module):
 
         # Create the model
         model = cls(**model_data)
+
         if finetune_config:
             # Apply the finetuning strategy
             model = apply_finetuning_strategy(model, finetune_config)
@@ -687,11 +688,14 @@ class PET(torch.nn.Module):
         model.to(dtype).load_state_dict(model_state_dict)
         model.additive_models[0].sync_tensor_maps()
 
+        # Loading the metadata from the checkpoint
+        metadata = checkpoint.get("metadata", None)
+        if metadata is not None:
+            model.__default_metadata__ = metadata
+
         return model
 
-    def export(
-        self, metadata: Optional[ModelMetadata] = None
-    ) -> MetatensorAtomisticModel:
+    def export(self, metadata: Optional[ModelMetadata] = None) -> AtomisticModel:
         dtype = next(self.parameters()).dtype
         if dtype not in self.__supported_dtypes__:
             raise ValueError(f"unsupported dtype {dtype} for PET")
@@ -723,11 +727,11 @@ class PET(torch.nn.Module):
         )
 
         if metadata is None:
-            metadata = ModelMetadata()
+            metadata = self.__default_metadata__
+        else:
+            metadata = merge_metadata(self.__default_metadata__, metadata)
 
-        append_metadata_references(metadata, self.__default_metadata__)
-
-        return MetatensorAtomisticModel(self.eval(), metadata, capabilities)
+        return AtomisticModel(self.eval(), metadata, capabilities)
 
     def _add_output(self, target_name: str, target_info: TargetInfo) -> None:
         # warn that, for Cartesian tensors, we assume that they are symmetric
@@ -762,35 +766,35 @@ class PET(torch.nn.Module):
             per_atom=True,
         )
 
-        self.node_heads[target_name] = nn.ModuleList(
+        self.node_heads[target_name] = torch.nn.ModuleList(
             [
-                nn.Sequential(
-                    nn.Linear(self.hypers["d_pet"], self.hypers["d_head"]),
-                    nn.SiLU(),
-                    nn.Linear(self.hypers["d_head"], self.hypers["d_head"]),
-                    nn.SiLU(),
+                torch.nn.Sequential(
+                    torch.nn.Linear(self.hypers["d_pet"], self.hypers["d_head"]),
+                    torch.nn.SiLU(),
+                    torch.nn.Linear(self.hypers["d_head"], self.hypers["d_head"]),
+                    torch.nn.SiLU(),
                 )
                 for _ in range(self.hypers["num_gnn_layers"])
             ]
         )
 
-        self.edge_heads[target_name] = nn.ModuleList(
+        self.edge_heads[target_name] = torch.nn.ModuleList(
             [
-                nn.Sequential(
-                    nn.Linear(self.hypers["d_pet"], self.hypers["d_head"]),
-                    nn.SiLU(),
-                    nn.Linear(self.hypers["d_head"], self.hypers["d_head"]),
-                    nn.SiLU(),
+                torch.nn.Sequential(
+                    torch.nn.Linear(self.hypers["d_pet"], self.hypers["d_head"]),
+                    torch.nn.SiLU(),
+                    torch.nn.Linear(self.hypers["d_head"], self.hypers["d_head"]),
+                    torch.nn.SiLU(),
                 )
                 for _ in range(self.hypers["num_gnn_layers"])
             ]
         )
 
-        self.node_last_layers[target_name] = nn.ModuleList(
+        self.node_last_layers[target_name] = torch.nn.ModuleList(
             [
-                nn.ModuleDict(
+                torch.nn.ModuleDict(
                     {
-                        key: nn.Linear(
+                        key: torch.nn.Linear(
                             self.hypers["d_head"],
                             prod(shape),
                             bias=True,
@@ -802,11 +806,11 @@ class PET(torch.nn.Module):
             ]
         )
 
-        self.edge_last_layers[target_name] = nn.ModuleList(
+        self.edge_last_layers[target_name] = torch.nn.ModuleList(
             [
-                nn.ModuleDict(
+                torch.nn.ModuleDict(
                     {
-                        key: nn.Linear(
+                        key: torch.nn.Linear(
                             self.hypers["d_head"],
                             prod(shape),
                             bias=True,