+++ tests/utils/test_output_gradient.py
@@ -1,7 +1,6 @@
-import metatensor.torch
 import pytest
 import torch
-from metatensor.torch.atomistic import System
+from metatomic.torch import System
 
 from metatrain.soap_bpnn import __model__
 from metatrain.utils.data import DatasetInfo, read_systems
@@ -106,7 +105,7 @@ def test_virial(is_training):
         for system in systems
     ]
     systems = [
-        metatensor.torch.atomistic.System(
+        System(
             positions=system.positions @ strain,
             cell=system.cell @ strain,
             types=system.types,
@@ -137,7 +136,7 @@ def test_virial(is_training):
         for system in systems
     ]
     systems = [
-        metatensor.torch.atomistic.System(
+        System(
             positions=system.positions @ strain,
             cell=system.cell @ strain,
             types=system.types,
@@ -190,7 +189,7 @@ def test_both(is_training):
         for system in systems
     ]
     systems = [
-        metatensor.torch.atomistic.System(
+        System(
             positions=system.positions @ strain,
             cell=system.cell @ strain,
             types=system.types,
@@ -219,7 +218,7 @@ def test_both(is_training):
         for system in systems
     ]
     systems = [
-        metatensor.torch.atomistic.System(
+        System(
             positions=system.positions @ strain,
             cell=system.cell @ strain,
             types=system.types,